Refactoring `sequence.py` into multiple files
