+++ pulser-core/pulser/sequence/_containers.py
@@ -1,0 +1,34 @@
+# Copyright 2022 Pulser Development Team
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Special containers used by the Sequence class."""
+from __future__ import annotations
+
+from collections import namedtuple
+from typing import NamedTuple, Union
+
+from pulser.pulse import Pulse
+from pulser.register.base_register import QubitId
+
+
+class _TimeSlot(NamedTuple):
+    """Auxiliary class to store the information in the schedule."""
+
+    type: Union[Pulse, str]
+    ti: int
+    tf: int
+    targets: set[QubitId]
+
+
+# Encodes a sequence building calls
+_Call = namedtuple("_Call", ["name", "args", "kwargs"])
